refactor: change logins button to button components by employee and company
