+++ src/views/auth/CompanyLogin.vue
@@ -6,7 +6,8 @@ import useNavigation from '@/composables/useNavigation.js';
 import apiLogin from '@/apis/user/login.js';
 import { setCookie } from '@/utils/cookie.js';
 import FormInput from '@/components/inputs/FormInput.vue';
-import Button from '@/components/buttons/BaseButton.vue';
+import Button from '@/components/Button.vue';
+// import Button from '@/components/buttons/BaseButton.vue';
 import LineDivider from '@/components/LineDivider.vue';
 import GoogleLogin from '@/views/auth/GoogleLogin.vue';
 
@@ -62,8 +63,9 @@ const submitLogin = async (value) => {
         placeholder="請輸入密碼"
       />
       <Button
+        size="full"
+        rounded="lg"
         role="company"
-        size="lg"
         :disabled="!meta.valid"
       >
         登入系統
