+++ src/views/auth/EmployeeLogin.vue
@@ -6,7 +6,7 @@ import useNavigation  from '@/composables/useNavigation.js';
 import { setCookie } from '@/utils/cookie.js';
 import apiLogin from '@/apis/user/login.js';
 import FormInput from '@/components/inputs/FormInput.vue';
-import Button from '@/components/buttons/BaseButton.vue';
+import Button from '@/components/Button.vue';
 
 const useProfile = useProfileStore();
 const { goToRoute } = useNavigation ();
@@ -62,8 +62,9 @@ const submitLogin = async (value) => {
         placeholder="請輸入密碼"
       />
       <Button
+        size="full"
+        rounded="lg"
         role="employee"
-        size="lg"
         :disabled="!meta.valid"
       >
         登入系統
